backend/pyproject.toml
[project]
name = "knowledge-agent"
version = "2.0.0"
description = "做深不做宽 — AI 深度知识萃取引擎"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "murphy"},
]

dependencies = [
    # LLM
    "zhipuai>=2.0.0",
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Data Fetching
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "aiohttp-socks>=0.8.0",
    # HTML Parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "readability-lxml>=0.8.1",
    # Config
    "ruamel.yaml>=0.18.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Date & Schedule
    "python-dateutil>=2.8.0",
    "APScheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
knowledge-agent = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

